Continue refactoring root signature cache

+++ engine/core/misc/hashes/fnv1a_64.h
@@ -14,7 +14,7 @@ public:
     void create(void const* p_data, size_t data_size) override;
     void combine(void const* p_data, size_t data_size) override;
     void finalize() override {}
-    uint8_t hashWidth() const override { return sizeof(m_hashData); }
+    size_t hashWidth() const override { return sizeof(m_hashData); }
     uint8_t const* hashValue() const override { return m_hashData.bytes; }
     std::strong_ordering operator<=>(HashValue const& other) const override
     {
